In computer class, the first assignment was to write a program to print the first 100 Fibonacci numbers. Instead, I wrote a program that would steal passwords of students

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -6,10 +6,9 @@
  */
 
 int main(void)
-
 {
 	int i = 0;
-	int j = 1, k = 2;
+	long j = 1, k = 2;
 
 	while (i < 50)
 	{
